Make the Website navigable for students
