+++ infrastructure/src/main/java/mops/controllers/LogoutController.java
@@ -1,0 +1,15 @@
+package mops.controllers;
+
+import javax.servlet.http.HttpServletRequest;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class LogoutController {
+
+  @GetMapping("/logout")
+  public String logout(HttpServletRequest request) throws Exception {
+    request.logout();
+    return "redirect:/";
+  }
+}
