+++ infrastructure/src/main/resources/templates/studenten/ergebnis.html
@@ -13,8 +13,8 @@
 <header>
     <nav class="navigation navigation-secondary" is="mops-navigation" th:fragment="navigation">
         <ul>
-            <li class="active"><a th:href="@{/feedback/}">Startseite</a></li>
-            <li><a th:href="@{/feedback/kontakt}">Contact us</a></li>
+            <li class="active"><a th:href="@{/feedback/student/}">Startseite</a></li>
+            <li><a th:href="@{/feedback/student/kontakt}">Contact us</a></li>
         </ul>
     </nav>
 </header>
