+++ infrastructure/src/main/resources/templates/studenten/kontakt.html
@@ -22,8 +22,8 @@
 <header>
     <nav class="navigation navigation-secondary" is="mops-navigation" th:fragment="navigation">
         <ul>
-            <li><a th:href="@{/feedback/}">Startseite</a></li>
-            <li class="active"><a th:href="@{/feedback/kontakt}">Contact us</a></li>
+            <li><a th:href="@{/feedback/student/}">Startseite</a></li>
+            <li class="active"><a th:href="@{/feedback/student/kontakt}">Contact us</a></li>
         </ul>
     </nav>
 </header>
