--- infrastructure/src/main/resources/templates/uebersicht.html
@@ -1,63 +1,0 @@
-<!DOCTYPE html>
-<html lang="en"
-      th:replace="~{mopslayout :: html(name='feedback', headcontent=~{:: headcontent}, navigation=~{:: navigation}, bodycontent=~{:: bodycontent})}"
-      xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="utf-8">
-    <title>feedback</title>
-    <th:block th:fragment="headcontent">
-        <!-- Links, Skripts, Styles hier einfügen! -->
-    </th:block>
-</head>
-<body>
-<header>
-    <nav class="navigation navigation-secondary" is="mops-navigation" th:fragment="navigation">
-        <ul>
-            <li class="active"><a th:href="@{/feedback/}">Startseite</a></li>
-            <li><a th:href="@{/feedback/kontakt}">Contact us</a></li>
-        </ul>
-    </nav>
-</header>
-<main th:fragment="bodycontent">
-    <h1 class="display-1">Dein Feedback</h1>
-    <div class="row row-cols-1 row-cols-md-4">
-        <div class="col mb-4" th:each="fragebogen :${frageboegen}">
-            <div class="card bg-light border-dark text-center">
-                <img alt="..." class="card-img-top" src="/uebung.jpg" height="150" weidth="225"
-                     th:if="${typeChecker.isUebung(fragebogen)}">
-                <img alt="..." class="card-img-top" src="/vorlesung.jpg " height="150" weidth="225"
-                     th:if="${typeChecker.isVorlesung(fragebogen)}">
-                <img alt="..." class="card-img-top" src="/aufgabe.jpg" height="150" weidth="225"
-                     th:if="${typeChecker.isAufgabe(fragebogen)}">
-                <img alt="..." class="card-img-top" src="/praktikum.jpg" height="150" weidth="225"
-                     th:if="${typeChecker.isPraktikum(fragebogen)}">
-                <img alt="..." class="card-img-top" src="/dozent.jpg" height="150" weidth="225"
-                     th:if="${typeChecker.isDozent(fragebogen)}">
-                <img alt="..." class="card-img-top" src="/gruppen.jpg" height="150" weidth="225"
-                     th:if="${typeChecker.isGruppe(fragebogen)}">
-                <img alt="..." class="card-img-top" src="/beratung.jpg" height="150" weidth="225"
-                     th:if="${typeChecker.isBeratung(fragebogen)}">
-                <div class="card-body">
-                    <h2 class="card-title" th:text="${fragebogen.veranstaltungsname}">Programmierung</h2>
-                    <p class="card-text">Verbleibende Zeit:</p>
-                    <table style="margin: auto">
-                        <tr class="clock_times">
-                            <td th:text="${fragebogen.getRemainingHours()}">--</td>
-                            <td th:text="${fragebogen.getRemainingMinutes()}">--</td>
-                            <td th:text="${fragebogen.getRemainingSeconds()}">--</td>
-                        </tr>
-                        <tr class="clock_names">
-                            <td>Hours</td>
-                            <td>Min.</td>
-                            <td>Sec.</td>
-                        </tr>
-                    </table>
-                    <a class="btn btn-primary" th:href="@{details(id=${fragebogen.bogennr})}">Feedback
-                        Geben!</a>
-                </div>
-            </div>
-        </div>
-    </div>
-</main>
-</body>
-</html>
