+++ infrastructure/src/test/java/mops/controllers/StudentControllerTest.java
@@ -17,7 +17,7 @@ import org.springframework.test.web.servlet.MockMvc;
 @AutoConfigureMockMvc
 @DisplayName("Tests for SubmitController")
 @SuppressWarnings( {"PMD.JUnitTestsShouldIncludeAssert", "PMD.JUnitAssertionsShouldIncludeMessage"})
-public class SubmitControllerTest {
+class StudentControllerTest {
   @Autowired
   private transient MockMvc mvc;
 
@@ -25,13 +25,13 @@ public class SubmitControllerTest {
   @DisplayName("Test to show that POST is allowed and you are redirected to /feedback")
   @Disabled
   public void submit() throws Exception {
-    mvc.perform(post("/feedback/details/submit/1")).andExpect(status().is3xxRedirection())
+    mvc.perform(post("/feedback/student/details/submit/1")).andExpect(status().is3xxRedirection())
         .andExpect(view().name("redirect:/feedback/"));
   }
 
   @Test
   @DisplayName("Test to show that GET is not allowed to avoid side effects")
   public void getIsNotAllowed() throws Exception {
-    mvc.perform(get("/feedback/details/submit/1")).andExpect(status().isMethodNotAllowed());
+    mvc.perform(get("/feedback/student/details/submit/1")).andExpect(status().isMethodNotAllowed());
   }
 }
